requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
psycopg2-binary==2.9.9
redis==5.0.1
pip~=22.0.2
amqp~=5.2.0
vine~=5.1.0
kombu~=5.4.2
billiard~=4.2.1
celery~=5.4.0
click~=8.1.7
python-dateutil~=2.9.0.post0
wcwidth~=0.2.13
six~=1.16.0
prompt_toolkit~=3.0.48
setuptools~=59.6.0